move views to dedicated view folder
