+++ client/src/views/DiscussionView.tsx
